fix(tests): move SQL injection tests in the jdbc module
As obvously those didn't make sense in Elastic
