+++ core/src/test/java/io/kestra/core/repositories/AbstractFlowRepositoryTest.java
@@ -189,60 +189,6 @@ public abstract class AbstractFlowRepositoryTest {
     }
 
     @Test
-    void labelFilterShouldResistSqlInjection() {
-        String tenant = TestsUtils.randomTenant(this.getClass().getSimpleName());
-
-        FlowWithSource flow = FlowWithSource.builder()
-            .id("sql-injection-test")
-            .namespace(TEST_NAMESPACE)
-            .tenantId(tenant)
-            .labels(Label.from(Map.of("mykey", "myvalue")))
-            .build();
-        flow = flowRepository.create(GenericFlow.of(flow));
-
-        try {
-            // EQUALS: injection in label key — payload must be treated as a literal string, returning no results
-            assertThat(
-                flowRepository.find(Pageable.UNPAGED, tenant, List.of(
-                    QueryFilter.builder()
-                        .field(Field.LABELS)
-                        .operation(Op.EQUALS)
-                        .value(Map.of("' OR '1'='1", "anything"))
-                        .build()
-                ))
-            ).as("EQUALS with SQL injection in label key should return no results").isEmpty();
-
-            // EQUALS: injection in label value — payload must be treated as a literal string, returning no results
-            assertThat(
-                flowRepository.find(Pageable.UNPAGED, tenant, List.of(
-                    QueryFilter.builder()
-                        .field(Field.LABELS)
-                        .operation(Op.EQUALS)
-                        .value(Map.of("mykey", "' OR '1'='1"))
-                        .build()
-                ))
-            ).as("EQUALS with SQL injection in label value should return no results").isEmpty();
-
-            // NOT_EQUALS: injection in label key — no flow has the injected key, so NOT_EQUALS must return the flow
-            // (tests the Postgres jsonb_path_query_first path and MySQL JSON_SEARCH path)
-            assertThat(
-                flowRepository.find(Pageable.UNPAGED, tenant, List.of(
-                    QueryFilter.builder()
-                        .field(Field.LABELS)
-                        .operation(Op.NOT_EQUALS)
-                        .value(Map.of("' OR '1'='1", "anything"))
-                        .build()
-                ))
-            ).as("NOT_EQUALS with SQL injection in label key should return the flow (literal comparison, key absent)")
-                .hasSize(1)
-                .extracting(Flow::getId)
-                .containsExactly("sql-injection-test");
-        } finally {
-            deleteFlow(flow);
-        }
-    }
-
-    @Test
     void shouldFilterFlowsWithNotEqualsLabelOperator() {
         String tenant = TestsUtils.randomTenant(this.getClass().getSimpleName());
