+++ jdbc-h2/src/test/java/io/kestra/repository/h2/H2SQLInjectionTest.java
@@ -1,0 +1,6 @@
+package io.kestra.repository.h2;
+
+import io.kestra.jdbc.repository.AbstractSQLInjectionTest;
+
+class H2SQLInjectionTest extends AbstractSQLInjectionTest {
+}
