+++ jdbc-mysql/src/test/java/io/kestra/repository/mysql/MysqlSQLInjectionTest.java
@@ -1,0 +1,6 @@
+package io.kestra.repository.mysql;
+
+import io.kestra.jdbc.repository.AbstractSQLInjectionTest;
+
+class MysqlSQLInjectionTest extends AbstractSQLInjectionTest {
+}
