+++ jdbc-postgres/src/test/java/io/kestra/repository/postgres/PostgresSQLInjectionTest.java
@@ -1,0 +1,6 @@
+package io.kestra.repository.postgres;
+
+import io.kestra.jdbc.repository.AbstractSQLInjectionTest;
+
+class PostgresSQLInjectionTest extends AbstractSQLInjectionTest {
+}
